Moved codec functions into its own file.
